Expose Platform.program for headless Elm programs
If you want to do command line things, like run elm-test on node, there
is no need to have dependencies on virtual-dom or anything. This
addition means you can initialize Elm programs without a UI, only ports
and effects. It looks like this:

    var app = Elm.MyThing.worker();

If you create a headless program, the ONLY way to initialize will be as
a worker. Using embed/fullscreen/… will not be possible.

This also resolves an issue (that I cannot find) where you could trick
Elm into giving you a value of type Never. That should not be possible
anymore!