Remove IIFEs from all "native" code, use new format
A value that appears in Elm as Native.List.foldr appears in JS as
_List_foldr. It starts with an underscore, which is not possible for
Elm variables, and it uses the module name after Native.* as the prefix.

Furthermore, when definition depends on other native code, it must be
wrapped in a function. For example Native.Time.now is now a function
that produces a task. That function is invoked in Elm code, ensuring
that all the dependencies are available independent of the ordering of
native code.