Revert addition of (=>) to Tuple library
I suspect there a bunch of folks happily using Elm that think more
operators is a bad move. I now have talked to a decent number who thing
(=>) in particular is pretty weak because (1) tuples are fine and easy
and (2) it implies a direction which does not really make sense.

When Haskell was doing their “burning bridges proposal” I thought it
was definitely not the right move, but as a happy user, I only see
those discussions peripherally and do not participate.

Point is, I’d rather go the more conservative route for now