There are many problems with window.Date so stop relying on it.
There are just a bunch of problems with relying on the JS Date
representation that I do not think can be solved with the JS
foundation. For example, the ISO 8601 parsing is impure for dates.

It appears that elm-community/elm-time provides similar functionality,
but in a pure way that takes the IANA time zone database into account.

The pure approach could be improved with something like
https://github.com/evancz/time-zone-database-proposal down the line.

Unclear if we should stick with this choice, but I would like to
explore alternatives first.